Updated header size and fastest times display size
Updated the height of the entries in the fastest times display to make it look cleaner.  Also updated the size of the header in the game to not be as clunky.  Because the header was updated, the smiley gifs also had to be resized. Finally, the level choice display was also updated to use the command attribute instead of button binding for button selection so that a tkinter error was not printed in the console.